Move lab specific models from sentry to clims
