+++ src/clims/models/activity_instance.py
@@ -9,12 +9,12 @@ from django.db.models.fields import TextField
 class ActivityInstance(Model):
     __core__ = True
 
-    sample = FlexibleForeignKey('sentry.Sample')
-    user_task = FlexibleForeignKey('sentry.UserTask')
+    sample = FlexibleForeignKey('clims.Sample')
+    user_task = FlexibleForeignKey('clims.UserTask')
 
     # Maps back to a workflow engine (e.g. Camunda)
-    external_id = TextField('sentry.external_id')
+    external_id = TextField()
 
     class Meta:
-        app_label = 'sentry'
-        db_table = 'sentry_activityinstance'
+        app_label = 'clims'
+        db_table = 'clims_activityinstance'
