+++ src/clims/models/result_file.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import, print_function
 from django.db import models
 from sentry.db.models import Model, FlexibleForeignKey
-from sentry.models.user_task import UserTask
+from clims.models.user_task import UserTask
 
 
 class ResultFile(Model):
@@ -16,5 +16,5 @@ class ResultFile(Model):
     user_task = FlexibleForeignKey(UserTask)
 
     class Meta:
-        app_label = 'sentry'
-        db_table = 'sentry_result_file'
+        app_label = 'clims'
+        db_table = 'clims_result_file'
