+++ src/sentry/api/endpoints/organization_member_issues_assigned.py
@@ -1,7 +1,8 @@
 from __future__ import absolute_import
 
 from sentry.api.bases import OrganizationIssuesEndpoint
-from sentry.models import Group, OrganizationMemberTeam, Team, UserTask
+from sentry.models import Group, OrganizationMemberTeam, Team
+from clims.models import UserTask
 from django.db.models import Q
 
 # TODO: Relation between UserTask and
