+++ src/sentry/api/endpoints/user_files.py
@@ -10,7 +10,7 @@ from sentry.api.base import Endpoint, SessionAuthentication
 from sentry.api.exceptions import ResourceDoesNotExist
 from sentry.api.serializers import serialize
 from sentry.models import ApiApplication, ApiApplicationStatus
-from sentry.models.sample import Sample
+from clims.models.sample import Sample
 from sentry.tasks.deletion import delete_api_application
 
 delete_logger = logging.getLogger('sentry.deletions.api')
