+++ src/sentry/api/endpoints/user_task_activity.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 from rest_framework.response import Response
 from sentry.api.bases.user_task import UserTaskBaseEndpoint
 from sentry.api.bases import OrganizationEndpoint
-from sentry.models import UserTask
+from clims.models import UserTask
 from sentry.models.activity import Activity
 from sentry.api.serializers import serialize
 from sentry.api.paginator import OffsetPaginator
